+++ lib/fraction.dart
@@ -1,5 +1,5 @@
-/// A package that helps you working with fractions, mixed fractions and
-/// egyptian fractions.
+/// A package that helps you work with fractions, mixed fractions and Egyptian
+/// fractions.
 library fraction;
 
 export 'src/extensions/fraction_num.dart';