+++ lib/src/extensions/fraction_string.dart
@@ -2,10 +2,13 @@ import 'package:fraction/fraction.dart';
 
 /// Extension method that adds [Fraction] functionalities to [String].
 extension FractionString on String {
-  /// Checks whether the string contains a valid representation of a fraction in
-  /// the 'a/b' format.
+  /// Checks whether the [String] either:
   ///
-  /// Note that 'a' and 'b' must be integers.
+  ///  - contains a valid representation of a fraction in the 'a/b' format;
+  ///  - contains a fraction glyph character.
+  ///
+  /// For example, this getter returns `true` if this [String] is `'1/4'` or
+  /// `'⅝'`.
   bool get isFraction {
     try {
       Fraction.fromString(this);
@@ -22,10 +25,10 @@ extension FractionString on String {
     }
   }
 
-  /// Converts the string into a [Fraction].
+  /// Converts this [String] into a [Fraction].
   ///
-  /// If you want to be sure that this method doesn't throw a
-  /// [FractionException], use `isFraction` before.
+  /// If you want to be sure that this method doesn't throw an exception, use
+  /// the [isFraction] getter before.
   Fraction toFraction() {
     try {
       return Fraction.fromString(this);